+++ changelog.md
@@ -7,6 +7,7 @@
 
 * `[calendar]` Added all events list window.
 * `[torch]` Added initial torch functionality (driver)
+* `[music player]` Added simple music player implementation
 
 ### Changed